fix re-export
