+++ controllers/contacts/index.js
@@ -1,0 +1,15 @@
+const favoriteChange = require("./favoriteChange");
+const createContact = require("./createContact");
+const deleteContact = require("./deleteContact");
+const updateContacts = require("./updateContacts");
+const getContact = require("./getContact");
+const getAllContacts = require("./getAllContacts");
+
+module.exports = {
+  favoriteChange,
+  createContact,
+  deleteContact,
+  updateContacts,
+  getContact,
+  getAllContacts,
+};
