+++ controllers/index.js
@@ -1,0 +1,15 @@
+const favoriteChange = require("./contacts/favoriteChange");
+const createContact = require("./contacts/createContact");
+const deleteContact = require("./contacts/deleteContact");
+const updateContacts = require("./contacts/updateContacts");
+const getContact = require("./contacts/getContact");
+const getAllContacts = require("./contacts/getAllContacts");
+
+module.exports = {
+  favoriteChange,
+  createContact,
+  deleteContact,
+  updateContacts,
+  getContact,
+  getAllContacts,
+};
